+++ conf.lua
@@ -1,9 +1,7 @@
--- Global constants
-require("src/global/constants")
-require("src/global/colors")
+local const = require("src/global/const")
 
 -- Set up initial stuffs
 function love.conf(t)
-    t.window.width = ROWS * TILE_SIZE
-    t.window.height = COLS * TILE_SIZE
+    t.window.width = const.GRID_ROWS * const.TILE_SIZE
+    t.window.height = const.GRID_COLS * const.TILE_SIZE
 end