+++ src/global/colors.lua
@@ -1,7 +1,11 @@
-_G.GRID_LINE_COLOR = { 0.2, 0.2, 0.2 }
-_G.BG_COLOR = { 0.1, 0.1, 0.1 }
-_G.TILE_COLOR = { 0.2, 0.2, 0.2 }
-_G.WALL_COLOR = { 0.2, 0.2, 0.2 }
-_G.SNAKE_COLOR = { 0.4, 0.4, 0.4 }
-_G.SNAKE_HEAD_COLOR = { 0.5, 0.5, 0.5 }
-_G.APPLE_COLOR = { 0.8, 0.8, 0.8 }
+local colors = {
+    GRID_LINE = { 0.2, 0.2, 0.2 },
+    BG = { 0.1, 0.1, 0.1 },
+    TILE = { 0.2, 0.2, 0.2 },
+    WALL = { 0.2, 0.2, 0.2 },
+    SNAKE = { 0.4, 0.4, 0.4 },
+    SNAKE_HEAD = { 0.5, 0.5, 0.5 },
+    APPLE = { 0.8, 0.8, 0.8 },
+}
+
+return colors