+++ src/global/const.lua
@@ -1,0 +1,10 @@
+local const = {
+    GAME_TITTLE = "Slither",
+    TILE_SIZE = 16,
+    GRID_ROWS = 32,
+    GRID_COLS = 32,
+    FONT_MAIN_SIZE = 32,
+    FONT_SUB_SIZE = 14
+}
+
+return const